+++ game/core/src/main/java/com/spacecomplexity/longboilife/game/GameScreen.java
@@ -14,16 +14,12 @@ import com.spacecomplexity.longboilife.Main;
 import com.spacecomplexity.longboilife.MainInputManager;
 import com.spacecomplexity.longboilife.game.building.Building;
 import com.spacecomplexity.longboilife.game.building.BuildingType;
-import com.spacecomplexity.longboilife.game.globals.Constants;
-import com.spacecomplexity.longboilife.game.globals.GameState;
-import com.spacecomplexity.longboilife.game.globals.MainCamera;
-import com.spacecomplexity.longboilife.game.globals.MainTimer;
+import com.spacecomplexity.longboilife.game.globals.*;
 import com.spacecomplexity.longboilife.game.tile.InvalidSaveMapException;
 import com.spacecomplexity.longboilife.game.tile.Tile;
 import com.spacecomplexity.longboilife.game.ui.UIManager;
 import com.spacecomplexity.longboilife.game.utils.*;
 import com.spacecomplexity.longboilife.game.world.World;
-import com.spacecomplexity.longboilife.game.globals.Filepaths;
 
 import java.io.FileNotFoundException;
 import java.util.Arrays;
@@ -103,6 +99,9 @@ public class GameScreen implements Screen {
 
         // Initialise the events performed from this script.
         initialiseEvents();
+
+        // Play the game soundtrack
+        Soundtrack.getSoundtrack().play();
     }
 
     /**
@@ -138,6 +137,9 @@ public class GameScreen implements Screen {
                 world.build(toBuild, mouse);
                 gameState.money -= cost;
 
+                // Play the build sound
+                new SoundEffect(Filepaths.BUILD_SOUND).play();
+
                 // Remove the selected building if it is wanted to do so
                 if (Arrays.stream(Constants.dontRemoveSelection)
                         .noneMatch(category -> gameState.placingBuilding.getCategory() == category)) {
@@ -155,6 +157,9 @@ public class GameScreen implements Screen {
                 // Build the building at the mouse location and charge the player accordingly
                 world.build(gameState.movingBuilding, mouse);
                 gameState.money -= cost;
+
+                // Play the build sound
+                new SoundEffect(Filepaths.BUILD_SOUND).play();
 
                 // Remove the old moving building and selected building
                 gameState.movingBuilding = null;
@@ -329,7 +334,7 @@ public class GameScreen implements Screen {
 
     @Override
     public void hide() {
-        
+
     }
 
     /**