+++ game/core/src/main/java/com/spacecomplexity/longboilife/game/ui/UIManager.java
@@ -14,10 +14,12 @@ import com.badlogic.gdx.utils.viewport.ScalingViewport;
 import com.badlogic.gdx.utils.viewport.Viewport;
 import com.spacecomplexity.longboilife.game.globals.Constants;
 import com.spacecomplexity.longboilife.game.globals.GameState;
+import com.spacecomplexity.longboilife.game.globals.Soundtrack;
 import com.spacecomplexity.longboilife.game.ui.game.*;
 import com.spacecomplexity.longboilife.game.ui.gameover.UIOverview;
 import com.spacecomplexity.longboilife.game.utils.EventHandler;
 import com.spacecomplexity.longboilife.game.globals.Filepaths;
+import com.spacecomplexity.longboilife.game.utils.SoundEffect;
 
 /**
  * Class to manage the UI in the game.
@@ -98,6 +100,10 @@ public class UIManager {
             uiElements = new UIElement[] {
                     new UIOverview(viewport, table, skin),
             };
+
+            // Pause the soundtrack and play the game over sound
+            Soundtrack.getSoundtrack().pause();
+            new SoundEffect(Filepaths.GAME_OVER_SOUND).play();
 
             return null;
         });