+++ .github/CHANGELOG.md
@@ -19,3 +19,4 @@
 18. [Game] Enhancement - Increased default window size to 1440x720 and changed map zoom factor accordingly
 19. [Global] Enhancement - Added actions to release snapshot and stable versions
 20. [Website] Enhancement - Refactor to point to our links, and automatically upload test coverage
+21. [Game] Enhancement - Added the soundtrack and sound effects